requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.6.1 